% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larval_dynamics_functions.R
\name{calc.L1}
\alias{calc.L1}
\title{calc.L1}
\usage{
calc.L1(
  beta,
  mf,
  mf.delay.in,
  expos,
  delta.vo,
  c.v,
  nuone,
  mu.v,
  a.v,
  expos.delay
)
}
\arguments{
\item{beta}{per blackfly biting rate on humans (product of proportion of blackfly bites take on humans (h) & reciprocal of the duration of the gonotrophic cycle (g))}

\item{mf}{vector with total number of mf in each person (extracted from main matrix, summed over each mf age cat per individual)}

\item{mf.delay.in}{vector of ??}

\item{expos}{vector with total exposure (age/sex-specific and individual) per individual in population}

\item{delta.vo}{density dependence when microfilarae tend to 0}

\item{c.v}{severity of constraining density-dep larval development (per dermal mf)}

\item{nuone}{per-capita development rate from L1 to L2 larvae (default 201.6 year-1)}

\item{mu.v}{per-capita mortality rate of blackfly vectors (default 26 year-1)}

\item{a.v}{per-capita microfilaria-induced mortality of blackfly vectors (default 0.39 year-1)}

\item{expos.delay}{vector of values from final (4th) column of matrix for exposure (to fly bites) (for L1 delay)exposure.delay) as each value for individual in population}
}
\value{
vector of mean L1 per individual
}
\description{
L1 (parasite life stages) dynamics in the fly population, assumed to be at equilibrium (modelled deterministically).
}
