% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larval_dynamics_functions.R
\name{calc.L2}
\alias{calc.L2}
\title{calc.L2}
\usage{
calc.L2(nuone, L1.in, mu.v, nutwo, mf, a.v, expos)
}
\arguments{
\item{nuone}{per-capita development rate from L1 to L2 larvae (default 201.6 year-1)}

\item{L1.in}{vector of in initial L1 or L1 from last timestep}

\item{mu.v}{per-capita mortality rate of blackfly vectors (default 26 year-1)}

\item{nutwo}{per-capita development rate from L2 to L3 larvae (default 201.6 year-1)}

\item{mf}{vector with delayed mf (?)}

\item{a.v}{per-capita microfilaria-induced mortality of blackfly vectors (default 0.39 year-1)}

\item{expos}{vector of values from final (4th) column of matrix for exposure (to fly bites) (for L1 delay)exposure.delay) as each value for individual in population}
}
\value{
vector of mean L2 per individual
}
\description{
L2 (parasite life stages) dynamics in the fly population, assumed to be at equilibrium (modelled deterministically).
delay of 4 days for parasites moving from L1 to L2
}
