% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larval_dynamics_functions.R
\name{calc.L3}
\alias{calc.L3}
\title{calc.L3}
\usage{
calc.L3(nutwo, L2.in, a.H, g, mu.v, sigma.L0)
}
\arguments{
\item{nutwo}{per-capita development rate from L2 to L3 larvae (default 201.6 year-1)}

\item{L2.in}{vector of L2 from main matrix for each individual}

\item{a.H}{proportion of L3 shed per bite (on any blood host); default is 0.8}

\item{g}{length of the gonotrophic cycle (0.0096)}

\item{mu.v}{per-capita mortality rate of blackfly vectors (default 26 year-1)}

\item{sigma.L0}{per-capita mortality rate of L3 in blackfly vectors}
}
\value{
vector of mean L3 per individual
}
\description{
L3 (parasite life stages) dynamics in the fly population, assumed to be at equilibrium (modelled deterministically).
}
