% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_worm_dynamics_functions.R
\name{change.worm.per.ind.treat}
\alias{change.worm.per.ind.treat}
\title{change.worm.per.ind.treat}
\usage{
change.worm.per.ind.treat(
  give.treat,
  iteration,
  treat.start,
  times.of.treat,
  treat.stop,
  onchosim.cov,
  treat.vec,
  DT,
  cum.infer,
  lam.m,
  phi,
  N,
  mort.fems,
  lambda.zero.in
)
}
\arguments{
\item{give.treat}{value (1 or 0) indicating whether treatment occurs}

\item{iteration}{iteration along vector of times}

\item{treat.start}{iteration where treatment starts}

\item{times.of.treat}{sequence of treatment times}

\item{treat.stop}{iteration where treatment stops}

\item{onchosim.cov}{vector of which individuals will be treated if treatment is given}

\item{treat.vec}{vector of times since treatment for each individual}

\item{DT}{timestep}

\item{cum.infer}{proportion of adult female worms made permanently infertility due to ivermectin at each round (lambda'_p in Hamley et al. 2019 supp); default is 0.345}

\item{lam.m}{embryostatic effects of ivermectin; lam.m is the max rate of treatment-induced sterility (lambda_max in Hamley et al. 2019 supp): default is 32.4 year-1}

\item{phi}{rate of decay of ivermectin-induced female worm sterlisation; default is 19.6 year-1}

\item{N}{total human population size}

\item{mort.fems}{vector of mortality rates (adult worms) for each age class/category; updated in change.worm.per.ind1}

\item{lambda.zero.in}{updated vector of per-capita rate that female worms lose their fertility & return to non-fertile state}
}
\value{
list containing vector of a) updated lambda values per individual, b) updated (or not) vector of times since treatment for each individual
}
\description{
updating worm compartments based on treatment (if this occurs) i.e., considering additional mortality and loss of fertility due to ivermectin treatment;
approach assumes individuals which are moved from fertile to non and fertile class due to treatment re-enter fertile class at standard rate
}
