% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_worm_dynamics_functions.R
\name{change.worm.per.ind2}
\alias{change.worm.per.ind2}
\title{change.worm.per.ind2}
\usage{
change.worm.per.ind2(
  DT,
  time.each.comp,
  compartment,
  new.worms.nf.fo,
  w.f.l.c,
  N,
  cur.Wm.nf,
  mort.fems,
  cur.Wm.f,
  omeg,
  male.tot.worms,
  worm.loss.males,
  lambda.zero.in,
  treat.vec
)
}
\arguments{
\item{DT}{timestep}

\item{time.each.comp}{Duration of each age class for adult worms (q_W in Hamley et al. 2019 supp); 1 year default}

\item{compartment}{which worm age-compartment (iteration k)}

\item{new.worms.nf.fo}{new non-fertile female worms (binomial draw)}

\item{w.f.l.c}{vector for worms coming from previous compartment (0 when k ==1)}

\item{N}{total human population size}

\item{cur.Wm.nf}{number of current non-fertile female worms per individual}

\item{mort.fems}{mortality value for females per individual (from change.worm.per.ind1 function)}

\item{cur.Wm.f}{number of current fertile female worms per individual}

\item{omeg}{omega value per individual (from change.worm.per.ind1 function)}

\item{male.tot.worms}{vector of total male worms per individual (from change.worm.per.ind1 function)}

\item{worm.loss.males}{vector of male worms lost per individual (from change.worm.per.ind1 function)}

\item{lambda.zero.in}{updated lambda values per individual (updated fertile to non fertile rate to account for treatment or non-updated) from change.worm.per.ind.treat}

\item{treat.vec}{updated (or not) vector of times since treatment for each individual from change.worm.per.ind.treat}
}
\value{
list containing vector of a) total male worms per individual (not updated on treatment, so from change.worm.per.ind1),
b) male worms lost per individual (not updated on treatment, so from change.worm.per.ind1), c) vector of non-fertile female worms per individual
d) vector of fertile female worms per individual, e) vector of number of non-fertile worms lost per individual,
f) vector of number fertile worms lost per individual, g) updated (or not) vector of times since treatment for each individual (taken from change.worm.per.ind.treat)
}
\description{
final function to calculate the change in the number of adult worms in one adult worm age class for all people
}
