% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_worm_dynamics_functions.R
\name{delta.h}
\alias{delta.h}
\title{delta.h}
\usage{
delta.h(delta.hz, delta.hinf, c.h, L3, m, beta, expos)
}
\arguments{
\item{delta.hz}{proportion of L3 larvae establishing/ developing to adult stage within human host, per bit, when ATP tends to 0}

\item{delta.hinf}{proportion of L3 larvae establishing/ developing to adult stage within human host, per bit, when ATP tends to infinity}

\item{c.h}{severity of transmission intensity - dependent parasite establishment within humans}

\item{L3}{mean number of L3 in fly population}

\item{m}{vector to human host ratio}

\item{beta}{per blackfly biting rate on humans (product of proportion of blackfly bites take on humans (h) & reciprocal of the duration of the gonotrophic cycle (g))}

\item{expos}{individuals total (age/sex-specific and individual-specific) exposure to blackfly bites}
}
\value{
vector of density-dependent (delta_h) values per individual in the population
}
\description{
proportion of L3 larvae (final life stage in the fly population) developing into adult worms in humans (Pi_H in Hamley et al. 2019)
}
