% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/larval_dynamics_functions.R
\name{delta.v}
\alias{delta.v}
\title{Density-dependence in vector}
\usage{
delta.v(delta.vo, c.v, mf, expos)
}
\arguments{
\item{delta.vo}{density dependence when microfilarae tend to 0}

\item{c.v}{severity of constraining density-dep larval development (per dermal mf)}

\item{mf}{vector with total number of mf in each person (extracted from main matrix, summed over each mf age cat per individual)}

\item{expos}{vector with total exposure (age/sex-specific and individual) per individual in population}
}
\value{
vector for density-dependent (delat_V) value for each individual
}
\description{
proportion of mf per mg developing into infective larvae within the vector
}
