% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_dynamics_functions.R
\name{derivmf.rest}
\alias{derivmf.rest}
\title{microfilariae age classes > 1 derivatives}
\usage{
derivmf.rest(mf.in, mf.mort, mf.move, mf.comp.minus.one, k.in)
}
\arguments{
\item{mf.in}{matrix of mf in > 1 age classes per individual.}

\item{mf.mort}{mf.mu rate specified for > 1 age classes in each human.}

\item{mf.move}{determines mf aging (moving rate to the next age class).}

\item{mf.comp.minus.one}{previous mf age class (compartment).}

\item{k.in}{previous k (for k1 = 0) for RDK4 method.}
}
\value{
vector for each individual
}
\description{
function called during RK4 for > 1 age classes of microfilariae (mf).
}
