% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_wrappers.R
\name{ep.equi.sim}
\alias{ep.equi.sim}
\title{Run EPIONCHO-IBM epidemiological model with or without interventions}
\usage{
ep.equi.sim(
  time.its,
  ABR,
  DT,
  treat.int,
  treat.prob,
  give.treat,
  treat.start,
  treat.stop,
  pnc,
  min.mont.age,
  delta.hz.in,
  delta.hinf.in,
  c.h.in,
  gam.dis.in
)
}
\arguments{
\item{time.its}{number of iterations (this input is a single value)}

\item{ABR}{annual biting rate (this input is a single value)}

\item{DT}{timestep (must be one day e.g., 1/366)}

\item{treat.int}{treatment interval in years e.g., 1 is every 1 year, 0.5 is every 6 months (this input is a single value)}

\item{treat.prob}{total population coverage (this input is a single value between 0 - 1)}

\item{give.treat}{takes 1 (MDA) or 0 (no MDA)}

\item{treat.start}{iteration where treatment starts}

\item{treat.stop}{iteration where treatment stops}

\item{pnc}{proportion of population which never receive treatment (single input value between 0 - 1)}

\item{min.mont.age}{minimum age for giving a skin snip (single input value, default is 5)}

\item{delta.hz.in}{this is a new user-input for density-dependence in humans (proportion of L3 larvae establishing/ developing to adult stage within human host, per bit, when ATP tends to 0)}

\item{delta.hinf.in}{this is a new user-input for density-dependence in humans (proportion of L3 larvae establishing/ developing to adult stage within human host, per bit, when ATP tends to infinity)}

\item{c.h.in}{this is a new user-input for density-dependence in humans (severity of transmission intensity - dependent parasite establishment within humans)}

\item{gam.dis.in}{this is a new user-input for individual-level exposure in humans (overdispersion parameter k_E determining degree of individual exposure heterogeneity; default value is 0.3)}
}
\description{
Runs the individual-based Onchocerciasis transmission model (EPIONCHO-IBM), based on Hamley et al. 2019
ep.equi.sim will run one repeat of the model, typically the mean of 500 repeats is required
model must be run to equilibrium (100 years), before treatment can begin
treatment is parameterised based on ivermectin
if the mf prevalence is zero 50 years after the final treatment, we assume elimination has occured
code is available which saves the equilibrium and receives it as an input
}
