% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intervention_functions.R
\name{os.cov}
\alias{os.cov}
\title{Treat individuals function}
\usage{
os.cov(all.dt, pncomp, covrg, N)
}
\arguments{
\item{all.dt}{matrix containing (among other things) information on compliance to treatment (whether or not an individual can be treated) and the age of each individual.}

\item{pncomp}{proportion non-compliers.}

\item{covrg}{total coverage of the population.}

\item{N}{human population size (default = 400).}
}
\value{
matrix of individuals to be treated (?)
}
\description{
function to calculate which individuals will be treated.
}
