% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_dynamics_functions.R
\name{prevalence.for.age}
\alias{prevalence.for.age}
\title{mf prevalence}
\usage{
prevalence.for.age(age, ss.in, main.dat)
}
\arguments{
\item{age}{minimum age for giving a skin snip (therefore age from which prevalence calculated)}

\item{ss.in}{takes mf per skin snip count object for each individual to convert to binary variable for prevalence}

\item{main.dat}{main matrix contain age of each individual (to ensure only calculate prevalence based o age from which skin snips are taken)}
}
\value{
value for prevalence
}
\description{
calculates mf prevalence in people based on a skin snip
}
