% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_dynamics_functions.R
\name{rotate}
\alias{rotate}
\title{rotate matrix}
\usage{
rotate(x)
}
\arguments{
\item{x}{matrix to rotate.}
}
\value{
rotated matrix (?)
}
\description{
function to rotate matrix, used in mf function.
}
