% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_worm_dynamics_functions.R
\name{Wplus1.rate}
\alias{Wplus1.rate}
\title{Wplus1.rate}
\usage{
Wplus1.rate(delta.hz, delta.hinf, c.h, L3, m, beta, expos, DT)
}
\arguments{
\item{delta.hz}{proportion of L3 larvae establishing/ developing to adult stage within human host, per bit, when ATP tends to 0}

\item{delta.hinf}{proportion of L3 larvae establishing/ developing to adult stage within human host, per bit, when ATP tends to infinity}

\item{c.h}{severity of transmission intensity - dependent parasite establishment within humans}

\item{L3}{mean number of L3 in fly population}

\item{m}{vector to human host ratio}

\item{beta}{per blackfly biting rate on humans (product of proportion of blackfly bites take on humans (h) & reciprocal of the duration of the gonotrophic cycle (g))}

\item{expos}{individuals total (age/sex-specific and individual-specific) exposure to blackfly bites}

\item{DT}{timestep}
}
\value{
vector of acquisition rates (probabilities) per individual in the population
}
\description{
Individual rate of acquisition of new infections (male and female adult worms) in humans
}
